Some improvements to proto representations of CONNECT and UNSUBACK.
`mqtt_proto::Connect` used a newtype `ConnectSessionExpiryInterval` for
its `session_expiry_interval` field, as opposed to `ConnAck` and `Disconnect`
that used `SessionExpiryInterval` directly. This was done because
CONNECT packets have a different interpretation of
the optional session expiry interval property than CONNACK and DISCONNECT -
CONNECT packets assume that the property being missing means
`Infinite` or `Duration(0)` in v3 (depending on `clean_start`) and
`Duration(0)` in v5, while CONNACK and DISCONNECT assume that the property
being missing means "use the original value in the CONNECT packet".
However there is an easier way to do this, by moving the default-ing behavior
to the encode and decode functions, which this change does. This defaulting
also allows the `session_expiry_interval` field to be moved to
the `ConnectOtherProperties` type.

`mqtt_proto::UnsubAck` put the `reason_codes` field in its `OtherProperties`
type because of the history of using `OtherProperties` for v5-specific fields.
However now `OtherProperties` is meant for default-able fields,
and the reason codes of an UNSUBACK are required in v5. This change moves
the field to `mqtt_proto::UnsubAck`.