Split `EventLoop` into `Disconnected` and `Connected` typestates.
`Disconnected::connect()` consumes the `Disconnected` and returns a `Connected`,
representing a connection that has received a successful CONNACK, as well as
the CONNACK itself.

`Disconnected::connect()` also returns a `DisconnectHandle` that can be used
to disconnect. This is a separate handle instead of being
`Connected::disconnect()` because user code will usually have the `Connected`
held up inside a `Connected::poll().await` call at the time that it decides
to disconnect.

The `Client::{connect, disconnect}` API are thus no longer longer necessary.
The disconnection is now only reported via `Connected::poll()`'s return type
instead of a completion token.