Move SharedImpl-specific types out from `Session` into `Client` and `Receiver`. (#42)
This commit changes the internal `*Request` types used in `Session` channels
to be generic on `S: Shared`. The external `AckHandle` and `*Token` types
are also made generic to handle the generic packets that they will use to
complete with. The `Client` and `Receiver` wrap the `AckHandle` and `*Token`s
in newtypes that parameterize the inner types on `SharedImpl` and
hide the `SharedImpl` from the user.

The main advantage here is that conversions from `Bytes` to `bytes::SharedImpl`
is a free no-op. `String` to `ByteStr<bytes::SharedImpl>` however
still requires a copy because `ByteStr` expects to have a length prefix;
this might be improved later by allowing `ByteStr` to hold values without
a length prefix. Nevertheless, the fact that the conversion is now infallible
and does not require an `Owned` allowed for the `IntoBuffered` trait to be
removed and its impls replaced with `From` impls.