Make `ConnectHandle` `Send`.
`TcpStream` types were wrapped in `Rc` because we didn't need them to be `Send`
in MQ, but we do need them to be `Send` here so wrap them in `Arc` instead.

The test streams and `Session` hold on to an `Owned` which also needs to be
`Send`, which is also not a problem for the current impls or for
the known future ones.